Extract mouse UI simulation command parsing
Move schema normalization and enum conversion into a dedicated internal command
type so subsequent validator extraction no longer depends on a nested private
type. Preserve every copied value, default, and validation message.
