+++ Packages/src/Editor/FirstPartyTools/SimulateMouseUi/MouseUiSimulationCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d72a04fed9944f7b9ad6228ef9b6b120
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
